bugfix(unix): Check for and ingore "Operating in progress"
Added some checking in the stderr reading of `run` in `service_unix` to
identify if the output ended with `Operation now in progress` is the
command was `launchctl`.

If the output has the suffix it is ignored and treated as a non-error,
all other standard error output is treaded as an error like normal.